other: Add release number to the doc
Replaced deploy action with an action called upon assigning a release number. Added release number to the header of the PDF and on the title page.

Moved glossary and scope to a different directory to prevent errors during PDF generation. Updated documentation-conventions to revise the sample outline with the revised outline of this doc